requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.5.3
typing-extensions==4.9.0

# LLM APIs
openai==1.7.2
anthropic==0.8.1
google-generativeai==0.3.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pypdf==3.17.4

# Data Processing
pandas==2.1.4
numpy==1.26.2
xmltodict==0.13.0

# CLI & Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Logging & Monitoring
loguru==0.7.2

# Validation
jsonschema==4.20.0
